sprint5-6 int -> float32

+++ internal/app/handlers/balance/get.go
@@ -9,7 +9,7 @@ import (
 
 type UserBalance struct {
 	Current   float32 `json:"current"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 func (h *BalanceHandler) GetHandler(
